+++ data-transfer-object/src/main/java/org/grnet/cat/dtos/assessment/registry/MetricNodeDto.java
@@ -1,17 +1,14 @@
 package org.grnet.cat.dtos.assessment.registry;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
-import org.grnet.cat.dtos.registry.template.Node;
 
 import java.util.List;
 
-@JsonPropertyOrder({ "id", "name", "type", "benchmark_value", "value", "result", "tests","label_algorithm_type","label_type_metric" })
+@JsonPropertyOrder({ "id", "name", "type", "benchmark_value", "label_algorithm_type","label_type_metric", "value", "result", "tests"})
 @Getter
 @Setter
 @AllArgsConstructor