+++ data-transfer-object/src/main/java/org/grnet/cat/dtos/registry/test/TestUpdateDto.java
@@ -10,16 +10,6 @@ public class TestUpdateDto {
     @Schema(
             type = SchemaType.STRING,
             implementation = String.class,
-            description = "The Test Name",
-            example = "T12"
-    )
-    @JsonProperty("tes")
-    @NotEmpty(message = "tes may not be empty.")
-    public String TES;
-
-    @Schema(
-            type = SchemaType.STRING,
-            implementation = String.class,
             description = "Label for the test",
             example = "PID Persistence - Service - Evidence"
     )