+++ mapper/src/main/java/org/grnet/cat/mappers/AssessmentMapper.java
@@ -3,17 +3,20 @@ package org.grnet.cat.mappers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.enterprise.inject.spi.CDI;
-import jdk.jshell.execution.Util;
 import lombok.SneakyThrows;
 import org.grnet.cat.dtos.assessment.AssessmentDoc;
 import org.grnet.cat.dtos.assessment.AdminPartialJsonAssessmentResponse;
 import org.grnet.cat.dtos.assessment.UserPartialJsonAssessmentResponse;
 import org.grnet.cat.dtos.assessment.registry.AdminJsonRegistryAssessmentResponse;
+import org.grnet.cat.dtos.assessment.registry.CriNodeDto;
 import org.grnet.cat.dtos.assessment.registry.RegistryAssessmentDto;
 import org.grnet.cat.dtos.assessment.registry.UserJsonRegistryAssessmentResponse;
 import org.grnet.cat.dtos.statistics.AssessmentPerActorDto;
 import org.grnet.cat.entities.AssessmentPerActor;
 import org.grnet.cat.entities.MotivationAssessment;
+import org.grnet.cat.repositories.registry.TestDefinitionRepository;
+import org.grnet.cat.repositories.registry.TestRepository;
+import org.grnet.cat.utils.TestParamsTransformer;
 import org.grnet.cat.utils.Utility;
 import org.mapstruct.IterableMapping;
 import org.mapstruct.Mapper;
@@ -23,7 +26,10 @@ import org.mapstruct.factory.Mappers;
 
 import java.sql.Timestamp;
 import java.time.Instant;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 /**
  * The AssessmentMapper is responsible for mapping Assessment entities to DTOs.
@@ -91,7 +97,6 @@ public interface AssessmentMapper {
     @Mapping(target = "sharedToUser", expression = "java(isSharedToUser(assessment.getValidation().getUser().getId()))")
     @Mapping(target = "sharedByUser", expression = "java(isRegistryAssessmentSharedByUser(assessment, assessment.getValidation().getUser().getId()))")
     @Mapping(target = "published", source = "assessment.published")
-
     UserJsonRegistryAssessmentResponse userRegistryAssessmentToJsonAssessment(MotivationAssessment assessment);
 
     @IterableMapping(qualifiedByName = "mapWithRegistryExpression")
@@ -150,10 +155,7 @@ public interface AssessmentMapper {
     @Mapping(target = "sharedToUser", ignore = true)
     @Mapping(target = "sharedByUser", ignore = true)
     @Mapping(target = "published", source = "assessment.published")
-
     UserJsonRegistryAssessmentResponse publicUserRegistryAssessmentToJsonAssessment(MotivationAssessment assessment);
-
-
 
     @Named("mapZenodoWithRegistryExpression")
     @Mapping(target = "assessmentDoc", expression = "java(registryStringJsonToDto(assessment.getAssessmentDoc()))")
@@ -165,7 +167,73 @@ public interface AssessmentMapper {
     @Mapping(target = "sharedToUser", expression = "java(!uniqueIdentifier.equals(assessment.getValidation().getUser().getId()))")
     @Mapping(target = "sharedByUser", expression = "java(assessment.getShared() && assessment.getValidation() != null && assessment.getValidation().getUser() != null && uniqueIdentifier.equals(assessment.getValidation().getUser().getId()))")
     @Mapping(target = "published", source = "assessment.published")
-
     UserJsonRegistryAssessmentResponse zenodoUserRegistryAssessmentToJsonAssessment(MotivationAssessment assessment,String  uniqueIdentifier);
 
+    @Named("jsonWithLatestChanges")
+    @Mapping(target = "assessmentDoc", expression = "java(assessmentWithLatestChanges(assessment))")
+    @Mapping(target = "validationId", expression = "java(assessment.getValidation().getId())")
+    @Mapping(target = "createdOn", expression = "java(assessment.getCreatedOn().toString())")
+    @Mapping(target = "userId", expression = "java(assessment.getValidation().getUser().getId())")
+    @Mapping(target = "updatedOn", expression = "java(assessment.getUpdatedOn() != null ? assessment.getUpdatedOn().toString() : \"\")")
+    @Mapping(target = "updatedBy", expression = "java(assessment.getUpdatedBy() != null ? assessment.getUpdatedBy() : \"\")")
+    @Mapping(target = "sharedToUser", expression = "java(isSharedToUser(assessment.getValidation().getUser().getId()))")
+    @Mapping(target = "sharedByUser", expression = "java(isRegistryAssessmentSharedByUser(assessment, assessment.getValidation().getUser().getId()))")
+    @Mapping(target = "published", source = "assessment.published")
+    UserJsonRegistryAssessmentResponse userRegistryAssessmentToJsonUpdatedWithLatestChangesAssessment(MotivationAssessment assessment);
+
+    @SneakyThrows
+    default RegistryAssessmentDto assessmentWithLatestChanges(MotivationAssessment assessment) {
+
+        var objectMapper = new ObjectMapper();
+
+        var testRepository = CDI.current().select(TestRepository.class).get();
+
+        var tests = testRepository.getUpdatedTests(assessment.getCreatedOn());
+
+        var json = objectMapper.readValue(assessment.getAssessmentDoc(), RegistryAssessmentDto.class);
+
+        if(!tests.isEmpty()){
+
+            var jsonTests = json
+                    .principles
+                    .stream()
+                    .flatMap(pri->pri.criteria.stream())
+                    .map(CriNodeDto::getMetric)
+                    .flatMap(metric->metric.getTests().stream())
+                    .collect(Collectors.toList());
+
+            jsonTests
+                    .forEach(jsonTest->{
+
+                        var optional =  tests
+                                .stream()
+                                .filter(obj -> obj.getTES().equals(jsonTest.getId()))
+                                .findFirst();
+
+                        if(optional.isPresent()){
+
+                            var dbTest = optional.get();
+
+                            jsonTest.setDescription(dbTest.getDescTest().trim());
+                            jsonTest.setName(dbTest.getLabelTest().trim());
+                            jsonTest.setResult(null);
+                            jsonTest.setValue(null);
+
+                            var testDefinitionRepository = CDI.current().select(TestDefinitionRepository.class).get();
+                            var testDefinition = testDefinitionRepository.fetchTestDefinitionByTestId(dbTest.getId());
+                            jsonTest.setParams(TestParamsTransformer.transformTestParams(testDefinition.getTestParams()));
+                            jsonTest.setToolTip(testDefinition.getToolTip().trim());
+                            jsonTest.setText(testDefinition.getTestQuestion().trim());
+                            jsonTest.setChanged(Boolean.TRUE);
+
+                            if(!Objects.isNull(jsonTest.getUrls())){
+
+                                jsonTest.setUrls(new ArrayList<>());
+                            }
+                        }
+                    });
+        }
+
+        return json;
+    }
 }