+++ service/src/main/java/org/grnet/cat/services/registry/TestService.java
@@ -6,14 +6,11 @@ import jakarta.transaction.Transactional;
 import jakarta.ws.rs.ForbiddenException;
 import jakarta.ws.rs.core.UriInfo;
 import org.grnet.cat.dtos.pagination.PageResource;
-import org.grnet.cat.dtos.registry.principle.PrincipleResponseDto;
 import org.grnet.cat.dtos.registry.test.*;
-import org.grnet.cat.entities.registry.Principle;
 import org.grnet.cat.entities.registry.Test;
 import org.grnet.cat.entities.registry.TestDefinition;
 import org.grnet.cat.exceptions.UniqueConstraintViolationException;
 import org.grnet.cat.mappers.registry.MotivationMapper;
-import org.grnet.cat.mappers.registry.PrincipleMapper;
 import org.grnet.cat.mappers.registry.TestMapper;
 import org.grnet.cat.repositories.registry.MetricTestRepository;
 import org.grnet.cat.repositories.registry.TestDefinitionRepository;
@@ -21,6 +18,8 @@ import org.grnet.cat.repositories.registry.TestRepository;
 
 import org.jboss.logging.Logger;
 
+import java.sql.Timestamp;
+import java.time.Instant;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
@@ -54,8 +53,6 @@ public class TestService {
         return TestMapper.INSTANCE.testToDto(test);
     }
 
-
-
     /**
      * Retrieves a specific Test item and related Test Definition item by its ID.
      *
@@ -69,9 +66,6 @@ public class TestService {
 
         return testResponseWithMotivations(test, testDefinition);
     }
-
-
-
 
     /**
      * This method takes a Principle entity, converts it to a PrincipleResponseDto, retrieves and maps
@@ -94,12 +88,10 @@ public class TestService {
         return testAndTestDefinitionResponse;
     }
 
-
-
     /**
      * Creates a new Test item.
      *
-     * @param userId         The user creating the Test.
+     * @param userId  The user creating the Test.
      * @param request The Test and Test Definition request data.
      * @return The created Test DTO.
      */
@@ -130,12 +122,13 @@ public class TestService {
     @Transactional
     public void updateTest(String id, String userId, TestAndTestDefinitionUpdateRequest request) {
 
-        if (metricTestRepository.existTestInStatus(id, Boolean.TRUE)) {
-            throw new ForbiddenException("No action permitted, test exists in a published motivation");
-        }
+//        if (metricTestRepository.existTestInStatus(id, Boolean.TRUE)) {
+//            throw new ForbiddenException("No action permitted, test exists in a published motivation");
+//        }
 
         var test = testRepository.findById(id);
         test.setPopulatedBy(userId);
+        test.setLastTouch(Timestamp.from(Instant.now()));
 
         if(!Objects.isNull(request.getTestRequest())){
 
@@ -147,7 +140,6 @@ public class TestService {
             var testDefinition = testDefinitionRepository.fetchTestDefinitionByTestId(id);
             testDefinitionService.updateTestDefinition(testDefinition.getId(), userId, request.getTestDefinitionRequest());
         }
-
     }
 
     /**
@@ -163,22 +155,6 @@ public class TestService {
 
         return testRepository.deleteById(id);
     }
-
-//    /**
-//     * Retrieves a page of Test items.
-//     *
-//     * @param page    The index of the page to retrieve (starting from 0).
-//     * @param size    The maximum number of Test items to include in a page.
-//     * @param uriInfo The Uri Info for pagination links.
-//     * @return A PageResource containing the Test items in the requested page.
-//     */
-//    public PageResource<TestResponseDto> getTestlistAll(int page, int size, UriInfo uriInfo) {
-//
-//        var testPage = testRepository.fetchTestByPage(page, size);
-//        var testDtos = TestMapper.INSTANCE.testToDtos(testPage.list());
-//
-//        return new PageResource<>(testPage, testDtos, uriInfo);
-//    }
 
     public PageResource<TestAndTestDefinitionResponse> getTestAndTestDefinitionListAll(String search, String sort, String order, int page, int size, UriInfo uriInfo) {