+++ CHANGELOG.md
@@ -23,6 +23,10 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 ## Unreleased
 ---
 
+### Added
+- [#475](https://github.com/FC4E-CAT/fc4e-cat-api/pull/475) CAT-859 Dynamically Update Tests When Fetching stored Assessment.
+
+
 ## 2.0.0 - 2025-03-31
 ---